+++ src/components/Contact/index.ts
@@ -1,1 +1,4 @@
+import Contact from "./Contact"
+import ContactForm from "./ContactForm"
 
+export { Contact, ContactForm }
