+++ src/components/index.ts
@@ -2,4 +2,5 @@ export { default as Home } from '@/components/Home/Home';
 export { default as Baklava } from '@/components/Baklava/Baklava';
 export { default as Locations } from '@/components/Locations/Locations';
 export { default as About } from '@/components/About/About';
+export { default as Contact } from '@/components/Contact/Contact';
 export { default as ContactForm } from '@/components/Contact/ContactForm';
