+++ src/styles/Contact.module.css
@@ -1,0 +1,17 @@
+/* CONTAINERS */
+.contact-content-container {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  padding: 20px;
+  background-color: #f9f9f9;
+}
+
+.contact-store-details-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 60%;
+  padding: 20px;
+  text-align: center;
+}
