+++ src/styles/globals.css
@@ -256,93 +256,7 @@ html {
   margin-top: 20px;
 }
 
-.contact-page-content {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  padding: 20px;
-  background-color: #f9f9f9;
-}
-
-.form-headings {
-  text-align: center;
-}
-
-.form-subheadings {
-  text-align: center;
-}
-
-fieldset {
-  margin-bottom: 16px;
-}
-
-.form-container {
-  max-width: 500px;
-  padding: 20px 100px;
-  border: 1px solid #eee;
-  border-radius: 10px;
-  background-color: #f9f9f9;
-  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-}
-
-.form-field-container {
-  margin-bottom: 30px;
-}
-
-.form-field {
-  width: 100%;
-  padding: 10px;
-  margin-bottom: 15px;
-  font-size: 16px;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-  box-sizing: border-box;
-  transition: border-color 0.2s ease, box-shadow 0.2s ease;
-}
-
-.form-field::placeholder {
-  color: #aaa;
-  font-style: italic;
-}
-
-.form-field:focus {
-  border-color: #007bff; /* Primary blue color for focus */
-  box-shadow: 0 0 5px rgba(0, 123, 255, 0.5);
-  outline: none;
-}
-
-.form-label {
-  display: block;
-  margin-bottom: 5px;
-  font-weight: bold;
-  font-size: 14px;
-  color: #333;
-}
-
-.form-error {
-  color: #d9534f; /* Red for errors */
-  font-size: 14px;
-  margin-top: -10px;
-  margin-bottom: 15px;
-}
-
-.form-radio {
-  margin-right: 10px;
-}
-
-.form-radio-label {
-  font-size: 14px;
-  color: #333;
-  cursor: pointer;
-}
-
-.error-border {
-  border: 2px solid red;
-  border-radius: 4px;
-  transition: border-color 0.3s ease;
-}
-
-.form-button {
+.button {
   display: inline-block;
   background-color: #007bff; /* Primary blue */
   color: #fff;
@@ -350,13 +264,4 @@ fieldset {
   padding: 10px 20px;
   border: none;
   border-radius: 5px;
-}
-
-.contact-store-details {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  height: 60%;
-  padding: 20px;
-  text-align: center;
 }
