+++ src/styles/index.ts
@@ -1,3 +1,4 @@
-import contactFormStyles from "./ContactForm.module.css";
+import formStyles from "./Form.module.css";
+import contactStyles from "./Contact.module.css";
 
-export { contactFormStyles }
+export { formStyles, contactStyles };
